pyproject.toml
[project]
name = "studybuddy-backend"
version = "0.1.0"
requires-python = ">=3.12.0"
dependencies = [
    "agno>=2.3.0,<2.3.5",
    "clerk-backend-api>=4.0.0",
    "fastapi>=0.115.6,<1.0.0",
    "ImageHash>=4.3.1,<5.0.0",
    "panoptodownloader",
    "Pillow>=11.0.0,<12.0.0",
    "psycopg2-binary>=2.9.10,<3.0.0",
    "pydantic>=2.10.0,<3.0.0",
    "python-dotenv>=1.0.1,<2.0.0",
    "python-multipart>=0.0.20,<1.0.0",
    "PyMuPDF>=1.24.11,<2.0.0",
    "requests>=2.32.3,<3.0.0",
    "sqlalchemy>=2.0.35,<3.0.0",
    "uvicorn>=0.34.0,<1.0.0",
    "google-genai>=1.52.0",
]

[tool.uv.sources]
panoptodownloader = { git = "https://github.com/novarii/Panopto-Video-DL-lib.git", rev = "relax-yarl" }
